Remove old graph.
