+++ crates/core/runner/src/lib.rs
@@ -1,11 +1,9 @@
 mod actions;
 mod dep_graph;
-mod dep_graph_new;
 mod errors;
 mod runner;
 mod subscribers;
 
-pub use dep_graph::DepGraph as DepGraphOld;
-pub use dep_graph_new::*;
+pub use dep_graph::*;
 pub use errors::*;
 pub use runner::*;
